+++ apps/its-battistar-be/src/api/app.router.ts
@@ -1,0 +1,9 @@
+import { Router } from 'express';
+
+import { todosRouter } from './todos/todos.router';
+
+const appRouter = Router();
+
+appRouter.use('/todos', todosRouter);
+
+export { appRouter };