+++ apps/its-battistar-be/src/db/redis/index.ts
@@ -1,0 +1,13 @@
+import IORedis from 'ioredis';
+
+import { e } from '../../environment';
+
+const connection = {
+  host: e.REDIS_HOST,
+  password: e.REDIS_PASSWORD,
+  port: Number(e.REDIS_PORT),
+  username: e.REDIS_USERNAME,
+  maxRetriesPerRequest: null,
+};
+
+export const redisConnection = new IORedis(connection);