+++ apps/its-battistar-be/src/types/express.d.ts
@@ -1,0 +1,6 @@
+declare namespace Express {
+  export interface Request {
+    requestTime: string | undefined;
+    user: null;
+  }
+}