+++ apps/its-battistar/src/app/pages/todos.service.ts
@@ -1,0 +1,8 @@
+import { Injectable } from '@angular/core';
+
+@Injectable({
+  providedIn: 'root',
+})
+export class TodosService {
+  message = 'Welcome to its-battistar!';
+}