+++ libs/shared-types/src/lib/index.ts
@@ -1,0 +1,4 @@
+export * from './entities/account.entity';
+export * from './entities/response.entity';
+export * from './entities/todo.entity';
+export * from './entities/user.entity';