+++ libs/shared-types/src/utils/ajv.ts
@@ -1,0 +1,8 @@
+import ajv from 'ajv';
+import addFormats from 'ajv-formats';
+
+const ajvInstance = new ajv();
+
+addFormats(ajvInstance);
+
+export default ajvInstance;