+++ lint-staged.config.js
@@ -1,8 +1,12 @@
 module.exports = {
-  '*.{css,scss,less,sass}': 'stylelint --fix --allow-empty-input',
+  '*.{css,scss,less,sass}':
+    'stylelint --fix --allow-empty-input && prettier --write',
+
+  '*.{ts,tsx,js,jsx,svelte,astro,cjs,mjs,cts,mts,vue}':
+    'eslint --cache --fix --ext .js,.ts,.tsx,.jsx,.svelte,.astro,.cjs,.mjs,.cts,.mts,.vue && prettier --write',
+
   '*.{yml,md,js,ts,tsx,svelte,astro,cjs,mjs,cts,mts,jsx,json,json5,.jsonc}':
     'prettier --write',
-  '*.{ts,tsx,js,jsx,svelte,astro,cjs,mjs,cts,mts,vue}':
-    'eslint --fix --ext .js,.ts,.tsx,.jsx,.svelte,.astro,.cjs,.mjs,.cts,.mts,.vue',
+
   // '*.{html}': 'html-validate && htmlhint',
 };