+++ apps/its-battistar/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
-This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
+This file was generated using
+[@jscutlery/semver](https://github.com/jscutlery/semver).
+
+## [0.1.0](https://github.com/gipo355/its-battistar/compare/its-battistar-0.0.1...its-battistar-0.1.0) (2024-4-21)
+
+### Features
+
+- **its-battistar:** :sparkles: prepare angular with shared lib for todos
+  component
+  ([30a2c50](https://github.com/gipo355/its-battistar/commit/30a2c50a10488cc7540a8f6b89b0ef46d533da7a))
 
 ## 0.0.1 (2024-4-9)